+++ packages/integration-sdk-core/src/errors.ts
@@ -290,6 +290,35 @@ export class IntegrationProviderAuthorizationError extends IntegrationProviderAP
   }
 }
 
+/**
+ * An error that may be thrown by integration during any step that interacts
+ * with provider APIs if the integration has exhausted all of the retries and can't fetch the resources.
+ *
+ * An alert SHOULD NOT be delivered to operators when this error is thrown to allow
+ * for improving the integration's handling of provider API errors.
+ */
+export class IntegrationProviderRetriesExceededError extends IntegrationProviderAPIError {
+  /**
+   * Optional encapsulation of the retry details received from API. Type is 'any' as the
+   * content will depend on the provider API.
+   */
+  readonly retryDetails?: any;
+
+  constructor(
+    options: IntegrationProviderApiErrorOptions & {},
+    retryDetails?: any,
+  ) {
+    super({
+      ...options,
+      code: 'PROVIDER_RETRIES_EXCEEDED_ERROR',
+      message: `Provider retries exceeded at ${options.endpoint}: ${options.status} ${options.statusText}`,
+      fatal: false,
+    });
+
+    this.retryDetails = retryDetails;
+  }
+}
+
 export type UserConfigError =
   | IntegrationValidationError
   | IntegrationProviderAuthenticationError;
@@ -308,7 +337,8 @@ export type ProviderAuthError =
 export function isProviderAuthError(err: Error): err is ProviderAuthError {
   return (
     err instanceof IntegrationProviderAuthorizationError ||
-    err instanceof IntegrationProviderAuthenticationError
+    err instanceof IntegrationProviderAuthenticationError ||
+    err instanceof IntegrationProviderRetriesExceededError
   );
 }