+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to
 
 ## Unreleased
 
+### Added
+
+- Additional error type `IntegrationProviderRetriesExceededError` to be used
+  when integration has exhausted all of the retries. This error type won't be
+  sent in as an alert to the operators.
+
 ## [8.10.1] - 2022-04-08
 
 ### Changed